refactor(spots): extract spots routes into dedicated blueprint and register; remove duplicates from customer
