+++ api/backend/rest_entry.py
@@ -9,6 +9,7 @@ from backend.simple.simple_routes import simple_routes
 from backend.ngos.ngo_routes import ngos
 from backend.o_and_m.o_and_m_routes import o_and_m
 from backend.customers.customer_routes import customer
+from backend.spots.spots_route import spots
 
 def create_app():
     app = Flask(__name__)
@@ -51,6 +52,7 @@ def create_app():
     app.register_blueprint(ngos, url_prefix="/ngo")
     app.register_blueprint(o_and_m, url_prefix="/o_and_m")
     app.register_blueprint(customer, url_prefix="/customer")
+    app.register_blueprint(spots, url_prefix="/spots")
 
     # Don't forget to return the app object
     return app
